Saving progress on bot testing refactor
NOTE: looks like I forgot the Screen class in the initial c++11 refactor

+++ src/xbs/Timer.h
@@ -12,7 +12,7 @@ namespace xbs
 {
 
 //-----------------------------------------------------------------------------
-typedef long long Milliseconds;
+typedef int64_t Milliseconds;
 typedef Milliseconds Timestamp;
 
 //-----------------------------------------------------------------------------
